Changed type to prevent error

+++ src/pkg/syscall/zerrors_dragonfly_amd64.go
@@ -4521,7 +4521,7 @@ const (
 // from http://golang.org/src/pkg/strconv/atoi.go
 intSize = 32 << uint(^uint(0)>>63)
 
-BIOCGBLEN = uint32(IOC_OUT | ((intSize & IOCPARM_MASK) << 16) | ((('B')) << 8) | ((102)))
+BIOCGBLEN = uintptr(IOC_OUT | ((intSize & IOCPARM_MASK) << 16) | ((('B')) << 8) | ((102)))
 
 //#define	_IOC(inout,group,num,len) \
 //	((unsigned long)(inout | ((len & IOCPARM_MASK) << 16) | ((group) << 8) | (num)))
